+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Export presets
   - Bundle settings for baking textures
   - Presets can be modified, added, and removed
+- Default export presets for
+  - Unreal Engine
 ### Changed
 - Setting the input texture size no longer overrides output texture size. The
   previous behavior was inconvenient and would cause users to constantly